Fixed a misunderstanding | See commit description
The first few versions of the library were published while still in development and critical breaking bugs and errors were found. Although v1.6 is the rewrite, it also had breaking errors. Version 1.7 was made to fix all these bugs and errors and actually be like DiscordEmojiHandler v1 can-be-used-in-production. 
The read me confusingly mentioned that all versions below 1.7 don't work, it meant DiscordEmojiHandler and not discord.py.
I reworded this at the end of the read me.

+ Also instead of removing the warning on the discord.py version, I fixed it to the actual version of DISCORD.PY that is needed which is 1.3. Though pretty much everyone is on discord.py 1.5+ so tell me if I should remove it entirely.